Update language switcher:
Do not use upper-case for locale selection.
Upper case is for country codes.
I think there was some confusion about those
- `NO` is Norway,
- `nb` is Norwegian Bokmål (as opposed to Nynorsk)
The Norwegian situation is rather complicated. Basically all Norwegians speak some dialect of the Norwegian language. Bokmål and Nynorks are two different _forms_ of that language. They are primarily distinct in written form, which of course is why we must choose one or the other when making an app.